# modularium_api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Modularium API",
    "version": "1.1.0",
    "description": "API for querying NFTs on Forma and Modularium marketplace data"
  },
  "paths": {
...
    "/collection/{tokenAddress}/{tokenId}": {
      "get": {
        "tags": [
          "Collection"
        ],
        "summary": "Get token",
        "description": "Returns a specific tokenId in the collection",
        "parameters": [
          {
            "$ref": "#/components/parameters/tokenAddressParam"
          },
          {
            "$ref": "#/components/parameters/tokenIdParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TokenRowResponse"
          },
          "400": {
            "$ref": "#/components/responses/InvalidTokenAddressError"
          }
        }
      }
    },
...
    "/collection/{tokenAddress}/stats": {
      "get": {
        "tags": [
          "Collection"
        ],
        "summary": "Get stats",
        "description": "Returns stats for a collection",
        "parameters": [
          {
            "$ref": "#/components/parameters/tokenAddressParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatsResponse"
          },
          "400": {
            "$ref": "#/components/responses/InvalidTokenAddressError"
          }
        }
      }
    },
    "/collection/{tokenAddress}/{tokenId}/stats": {
      "get": {
        "tags": [
          "Collection"
        ],
        "summary": "Get token stats",
        "description": "Returns stats for a specific tokenId in the collection",
        "parameters": [
          {
            "$ref": "#/components/parameters/tokenAddressParam"
          },
          {
            "$ref": "#/components/parameters/tokenIdParam"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatsResponse"
          },
          "400": {
            "$ref": "#/components/responses/InvalidTokenAddressError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "tokenAddressParam": {
        "in": "path",
        "name": "tokenAddress",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^0x[a-fA-F0-9]{40}$"
        },
        "description": "Address of the token contract",
        "example": "0xabcdef0123456789abcdef0123456789abcdef01"
      },
      "tokenIdParam": {
        "in": "path",
        "name": "tokenId",
        "required": true,
        "schema": {
          "type": "integer"
        },
        "description": "ID of the specific token",
        "example": 1
      },
...
    },
    "responses": {
      "InvalidTokenAddressError": {
        "description": "Invalid token address",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "invalidTokenAddress": {
                "$ref": "#/components/examples/invalidTokenAddressError"
              }
            }
          }
        }
      },
...
      "TokenRowResponse": {
        "description": "A single token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TokenRow"
            }
          }
        }
      },
...
      "StatsResponse": {
        "description": "Stats for a collection",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Stats"
            }
          }
        }
      },
...
    },
    "examples": {
...
      "invalidTokenAddressError": {
        "value": {
          "error": "Invalid token address"
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "Stats": {
        "type": "object",
        "properties": {
          "numMinted": {
            "type": "number",
            "description": "Number of tokens minted"
          },
          "totalSupply": {
            "type": "number",
            "description": "Total supply"
          },
          "primaryVolume": {
            "type": "number",
            "description": "Primary volume"
          },
          "totalSalesQty": {
            "type": "number",
            "description": "Total sales quantity"
          },
          "totalSalesVolume": {
            "type": "number",
            "description": "Total sales volume"
          },
          "maxSalePrice": {
            "type": "number",
            "description": "Maximum sale price"
          },
          "minSalePrice": {
            "type": "number",
            "description": "Minimum sale price"
          },
          "avgSalePrice": {
            "type": "number",
            "description": "Average sale price"
          },
          "floorPrice": {
            "type": "number",
            "description": "Floor price"
          },
          "numListed": {
            "type": "number",
            "description": "Number of listings"
          },
          "bestOffer": {
            "type": "number",
            "description": "Best offer"
          },
          "numOffers": {
            "type": "number",
            "description": "Number of offers"
          },
          "holders": {
            "type": "number",
            "description": "Number of holders"
          }
        }
      },
...
      "Metadata": {
        "type": "object",
        "additionalProperties": true,
        "description": "Metadata for the token",
        "example": {
          "name": "Mammoth #1",
          "description": "It's widely believed that the last woolly mammoths went extinct 9,999 years ago. In fact, they've continued to live, unseen, in the background of human history, bearing witness to its horrors, its triumphs, and its folly. They were present at Golgotha. They were there on the grassy knoll. They've watched the toppling of tyrants and the fast-changing fortunes of empires. Quietly, mammoths observe, withholding judgment, wanting nothing other than to fit in.",
          "image": "ipfs://bafybeihmfp45w4p3bk7vzekixhxsdrsyvtgs7gtiaestdcwoxrsts36e6u",
          "tags": [
            "modularium",
            "mammoth"
          ]
        }
      },
...
      "TokenType": {
        "type": "string",
        "enum": [
          "ERC721",
          "ERC1155"
        ],
        "description": "Type of the token"
      },
      "TokenRow": {
        "type": "object",
        "properties": {
          "tokenAddress": {
            "type": "string",
            "pattern": "^0x[a-fA-F0-9]{40}$",
            "description": "Contract address of the token",
            "example": "0xabcdef0123456789abcdef0123456789abcdef01"
          },
          "tokenId": {
            "type": "string",
            "description": "ID of the token",
            "example": "1"
          },
          "tokenType": {
            "$ref": "#/components/schemas/TokenType"
          },
          "totalSupply": {
            "type": "string",
            "description": "Total supply of the token",
            "example": "1"
          },
          "isMarketplaceAllowed": {
            "type": "boolean",
            "description": "Whether the token is allowed to be traded on the Modularium marketplace",
            "example": true
          },
          "tokenMetadata": {
            "$ref": "#/components/schemas/Metadata"
          }
        }
      },
...
    }
  },
...
}